#2 CT MR data LR flipped - fixed
The problem was because we need to reverse data along the z-axis to account for the different coordinate systems between Unity and VTK. This flipping was only being done if the data were unsigned char, e.g. US3D, and not for other data types, e.g. CT, MR (unsigned) short.

Also dicom data may not have been loaded unless they ended with a '.dcm' - which many dicom files don't. So assume that if the first file in the folder has no extension we are attempting to load in dicom data.