#8 VtkToUnity Binaries - updated instructions
No other significant change.

I wanted to make the submodule pull a specific tag, but there is no way to do this in git so far as I can tell. So the fix will be to manually switch to the appropriate build if we need to rollback in the future.